requirements.txt
anthropic>=0.34.0
python-pptx>=0.6.23
pyyaml>=6.0.1
streamlit>=1.28.0

